04_rag_vector_db/notebooks/requirements.txt
langchain-aws==0.2.22
langchain-community==0.3.23
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langgraph==0.4.1
faiss-cpu>=1.7,<2
opensearch-py==2.8.0
docling==2.31.0
pypdf==5.4.0
rank_bm25==0.2.2

